Added static method to DetermineRankingFromScores. Updated TrumbusTrace and SubsceneStarter.

+++ Assets/Game/1 Examples/TrumbusTrace/Misc/Stencil.cs
@@ -49,7 +49,7 @@ namespace Examples.TrumbusTrace {
             }
         }
 
-        public float CalculateScore() {
+        public int CalculateScore() {
             float score = 0;
             foreach (KeyValuePair<Vector3, float> entry in _pointDistances) {
                 if (entry.Value < scoreDistanceThreshold) {
